requirements.txt
flask
python-dotenv
langchain
langchain-huggingface
langchain-community
pypdf
pyjwt
faiss-cpu
pymongo[srv]
Flask-PyMongo
flask_cors